requirements.txt
django==1.11.4
pytz==2017.2
psycopg2==2.7.3
djangorestframework==3.6.3
boto3==1.4.6
botocore==1.6.3
docutils==0.14
jmespath==0.9.3
python-dateutil==2.6.1
s3transfer==0.1.10
six==1.10.0
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
kombu==4.1.0
vine==1.1.4
Pillow==4.3.0
olefile==0.44
numpy==1.13.3
decorator==4.0.11
imageio==2.1.2
moviepy==0.2.3.2
tqdm==4.11.2
python-magic==0.4.13
certifi==2017.11.5
chardet==3.0.4
idna==2.6
requests==2.18.4
stripe==1.75.0
urllib3==1.22
hashids==1.2.0
fcm-django==0.2.13
pep8==1.7.1
pycodestyle==2.3.1
